tapdb/sqlc/migrations: add new migration 38 for the universe supply tree
In this commit, we add a new migration for the universe supply trees.
Rather than use the multi-verse pointers (which are namespaced by proof
type), we instead move to use a new set of smt tree pointers, that are
scoped by the group key instead.

We'll use these later to keep track of the new supply sub-trees, and
also root supply tree. We also add a new proof type of mint_supply, to
distinguish from the normal issuance type.